# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/device/create-device": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает новое устройство и возвращает его данные",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Создать новое устройство",
                "parameters": [
                    {
                        "description": "Данные нового устройства",
                        "name": "device",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/serializer.DeviceJSON"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Созданное устройство",
                        "schema": {
                            "$ref": "#/definitions/serializer.DeviceJSON"
                        }
                    },
                    "400": {
                        "description": "Неверные данные",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/device/{id}": {
            "get": {
                "description": "Возвращает информацию об устройстве по её идентификатору",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Получить устройство по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID устройства",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные устройства",
                        "schema": {
                            "$ref": "#/definitions/serializer.DeviceJSON"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Устройство не найдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/device/{id}/add-photo": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Загружает изображение для устройства и возвращает обновленные данные",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Загрузить изображение устройства",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID устройства",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Изображение устройства",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Статус загрузки и данные устройства",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Неверный запрос или файл",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Устройство не найдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/device/{id}/add-to-amperage_application": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Добавляет устройство в заявку-черновик пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Добавить устройство в расчёт",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID устройства",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Расчёт с добавленным устройством",
                        "schema": {
                            "$ref": "#/definitions/serializer.AmperageApplicationJSON"
                        }
                    },
                    "201": {
                        "description": "Создан новый расчёт",
                        "schema": {
                            "$ref": "#/definitions/serializer.AmperageApplicationJSON"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Устройство не найдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Устройство уже в расчёте",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/device/{id}/delete-device": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Выполняет логическое удаление устройство по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Удалить устройство",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID устройства",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Статус удаления",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Устройство не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/device/{id}/edit-device": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновляет информацию об устройстве по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Изменить данные устройства",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID устройства",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные устройства",
                        "name": "device",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/serializer.DeviceJSON"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленное устройство",
                        "schema": {
                            "$ref": "#/definitions/serializer.DeviceJSON"
                        }
                    },
                    "400": {
                        "description": "Неверные данные",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Устройство не найдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/devices": {
            "get": {
                "description": "Возвращает все устройства или фильтрует по названию",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Получить список устройств",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название устройства для поиска",
                        "name": "device_title",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список устройств",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/serializer.DeviceJSON"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "serializer.AmperageApplicationJSON": {
            "type": "object",
            "properties": {
                "amperage": {
                    "type": "number"
                },
                "amperage_application_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "creator_login": {
                    "type": "string"
                },
                "finish_date": {
                    "type": "string"
                },
                "form_date": {
                    "type": "string"
                },
                "moderator_login": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "serializer.DeviceJSON": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "dev_power": {
                    "type": "number"
                },
                "device_id": {
                    "type": "integer"
                },
                "is_delete": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    }
}